Revert "Move CrashlyticsWorker to concurrency package (#6148)"
This reverts commit 99a9d9358502f88f464484c30539f836aa56e05b.
